Finished week7 assignment as well as primary functionality for the file-server. Basically it sucks but kind of works so whoo
