--- final_project/client/home/school_doc.txt.downloaded
@@ -1,1 +1,0 @@
-This is a school document.
