--- final_project/client/keys/newuser/newuser_priv.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEowIBAAKCAQEAqhZH8atf1B6yBGnmKb4ZQwTB8j4ZvMq9tSZnld/Fgm1NIxYC
-7k0fUr5vcXBljKWrwJaeY3U1S5+L+MEYdLAy0AVq6j3mO6kOfIrvmpbHgpxEifNS
-8edpZzMNGmKIGR33YU8QFMLkuCTjDvdIYvmXmSJER8J//p8e1w+ADlLdQLLh04R7
-uMkZvvWWtqgIBNnDnSPgeHbvHhpbEDz+P3SJQT8T9ZPEj4bhdwyY/mY84r3qrwR1
-u/dibjwO3TjHAHc9BbkmKwiPE/nHyb+wlmk9GCuCeekemkoKGnI6F+lbCIOHtFXn
-6QmjQJkrYAnVl8NbUBoPxg1nAPgb59k0xehDMwIDAQABAoIBACBkR+VFbuWVxu9C
-P7wZYPQ7drgsjXg/4CscmkOXrySsrjENYxeBBoMak1y2PBCyCGiQBrAIRVMWmnCQ
-TZkDaVnmfjsH+M3J32FmkEUPjAVav07KUTSSjqdBuLAarbKi21bj2LsrdqV/GOLj
-wxxeTjmypfBIoVmbciMg55FSxKIqX47JZ6YlMjoEzyRxXiPMDbYqD32Q2I1epgMN
-b9snSgADA4hMIhZIB4HPIhcB85zXp+qfU74gC7zCLL6yUUdGsI7qU/bWB9JFUraO
-3DcXwXdoIE9bLEJtqlEkawjsQMQvQe9G5akYraSVYu+WRhp/r/fOztVdl6WNEQqK
-eG9k6bkCgYEA0Itxmvo3nHJ0FCnYyD2ECNgGq8/lN8ZsVQs4STSWOhGoQbXY/A8R
-MyfN5DCz6rg590kZ4d3LCX3Q4BguCyxKXNngtsJfuJ82EhUbx+avbPLl/A3AfXWm
-PVTGKQB3GQ6dsghMZZ4ji0/HF3jCMzHhExQ6KPiYyHDTQbrAzrnBJocCgYEA0MqG
-jVBtLVJE0GfBg+SUcTuh4QrFEuEkPfrwLWkSwNUIMicfgXaM52U9hMiw5Jl+83Q1
-JhUwRJNXPEoI/aMSey6tpr6fOZnYginZkiFWfInoY7+rmekt8qkGATA9CZ976ZkR
-WHMtAYmBVckvhQKX9emRXyNY/ZahoU6EnbezfPUCgYB32nxbJk0Js9m0sOwvAPO+
-JXGIEv5YlMg0BDPAFuHz06wQv9NJ4drvvIOfDsKINoLR9QhUACUtekquUBc+QqEm
-Re3Y+1MKBy6LK+iAdbsCcYLt3nVvZUs6WltcuzC3jOfiptRJXp5s9VelzLhHtfX2
-Ahz+6QjdwcIYN52R8b3pRwKBgQC5Daev/Ean0rDXHf7wTVob+kSB1y/Qiee7X3wc
-Y3KRor7+5JoR/ubn+F1MISxXkIxqbubkeAct4IGDGANb+nrY3wheL7mkLe0tgLYr
-xIyJGRn1vt3zLi7hUlnWF9EhhE8pPdeX041iPWN6izQo6aGR/ZAxCH1JVYawF42G
-1FirOQKBgBBKZsQS9H5AdLkXWWLzgqx2CuNbFHV/q9xzXZON5pVBuxeHLHyJmgrv
-nSkwGUOW0UV5h5MbfmDW0Tex/djuRph9aDYe6LWP1HjkLkB7UECuDhc23ees2nSc
-VgnA+5v6t5c17WIP3y2qdkmDsjsu8zxjQPxJAPA80QGmJYl/HeXo
------END RSA PRIVATE KEY-----
