--- final_project/client/keys/newuser/newuser_pub.pem
@@ -1,9 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAqhZH8atf1B6yBGnmKb4Z
-QwTB8j4ZvMq9tSZnld/Fgm1NIxYC7k0fUr5vcXBljKWrwJaeY3U1S5+L+MEYdLAy
-0AVq6j3mO6kOfIrvmpbHgpxEifNS8edpZzMNGmKIGR33YU8QFMLkuCTjDvdIYvmX
-mSJER8J//p8e1w+ADlLdQLLh04R7uMkZvvWWtqgIBNnDnSPgeHbvHhpbEDz+P3SJ
-QT8T9ZPEj4bhdwyY/mY84r3qrwR1u/dibjwO3TjHAHc9BbkmKwiPE/nHyb+wlmk9
-GCuCeekemkoKGnI6F+lbCIOHtFXn6QmjQJkrYAnVl8NbUBoPxg1nAPgb59k0xehD
-MwIDAQAB
------END PUBLIC KEY-----
