--- final_project/client/keys/newuser/newuser_sym.key
@@ -1,1 +1,0 @@
-bbYMWlEBJGMO9CVAP0urWu_vVb_kHOFQfyMmWT3hnhg=
