--- final_project/server/folder_clients/dforkey/school_doc.txt
@@ -1,1 +1,0 @@
-gAAAAABhZ6xMyO0REnb1FBYUdiPJePM3h11GunLbat5Sa55VMe-W5zeoNcNNSBvKGk1z8Ee3-zBsXm5-E3VLQKQn28ceEwbt9hgEjDaXhRdk6Xx6UKLb9K8=
