+++ final_project/server/folder_clients/spallas/school_doc.txt
@@ -1,0 +1,1 @@
+gAAAAABhad_FZ05lSOK9CA2QaPkU4aJ-u5vP5IGap69vnr10FttExPXDrjwmJfUdnhO5z9shG9OCVknyUFmjPtQxJc5gZdvjoLk-m2SJTPXb2N1Oyf6ndj8=
